Add both a zipped and unzipped version of the Persona Template.
I'm still on the fence about which format (zipped or unzipped)
should exist in the repo.  For now I've added both.  I'd like to
treat it more like a binary file than a directory, but the
directory format may make it easier for others to just open it,
edit and contribute.

Signed-off-by: Tim Shadel <github@timshadel.com>